Конышева 25-ИСТ-1-1 alg.cpp

+++ src/alg.cpp
@@ -24,7 +24,7 @@ unsigned int collatzLen(uint64_t num) {
 }
 
 unsigned int seqCollatz(unsigned int* maxlen, uint64_t lbound, uint64_t rbound) {
-    unsigned int num = 0;
+    uint64_t num = 0;
     unsigned int resmax = 0;
 
     for (uint64_t i = lbound; i <= rbound; i++) {
